Huffman-Coding/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
python-dotenv==0.19.0
Flask-Mail==0.9.1
itsdangerous==2.0.1
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.47.0 